# perf: improve FTS performance for long query