this improves WAND perf in case of the query consists of many tokens:
- use bubble sort when the algo only needs to advance single iterator
- move less posting iterators when evaluate the documents if there are many iterators are at the same doc id