@BubbleCal 2025-08-27
tested with a query consisting of 50 tokens, this PR improves the perf by 3x~4x:

before:
```
FTS WAND performance (long query): 343.58ms, plan:
AnalyzeExec verbose=true, metrics=[]
  TracedExec, metrics=[]
    GlobalLimitExec: skip=0, fetch=10, metrics=[output_rows=10, elapsed_compute=1.542µs]
      MatchQuery: query=struct table library data signal query list language server parallel source software document type document socket text parallel table integer queue queue method time value analysis platform queue device compile vector response node service file resource analysis format device package trace encode package release perform deploy route update device load, metrics=[output_rows=10, elapsed_compute=343.155583ms, index_comparisons=56, indices_loaded=0, parts_loaded=0]
```

after:
```
FTS WAND performance (long query): 79.76ms, plan:
AnalyzeExec verbose=true, metrics=[]
  TracedExec, metrics=[]
    GlobalLimitExec: skip=0, fetch=10, metrics=[output_rows=10, elapsed_compute=1.542µs]
      MatchQuery: query=scale method network graph pointer develop interface register pointer hash value digital register device execute hardware runtime binary execute vector remote function request filter optimize document record token response pattern vector integer network record deploy filter storage function token binary update integer database file compile binary binary deploy list response, metrics=[output_rows=10, elapsed_compute=79.346708ms, index_comparisons=56, indices_loaded=0, parts_loaded=0]
```