@BubbleCal 2025-08-27
this also improves flat search perf by ~50%